Refactor of the entire code base. Included the proxy into the repo and changed the docker compose file. Did also implement celery to handle background tasks.
