+++ .gitignore
@@ -1,1 +1,2 @@
-/api.env
+/server/api.env
+/.venv
