--- enums/status.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class AudioStatus(Enum):
-    Complete = "COMPLETE"
-    Processing = "PROCESSING"
-
