+++ proxy/requirements.txt
@@ -1,0 +1,3 @@
+Flask
+mitmproxy
+requests
