+++ server/enums/status.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class AudioStatus(Enum):
+    Complete = "COMPLETE"
+    Processing = "PROCESSING"
+
