--- tests/pytest.ini.py
@@ -1,3 +1,0 @@
-[pytest]
-log_cli = true
-log_cli_level = INFO
